FE page Checkout, Shopping Cart

+++ src/main/java/ued/OrganicWeb/controller/web/api/OrderCartAPI.java
@@ -55,19 +55,29 @@ public class OrderCartAPI extends HttpServlet{
 		UserModel userModel =(UserModel) SessionUtil.getInstance().getValue(req, "user");
 		CustomerModel model = null;
 		Map<Integer, Integer> listItems = (HashMap<Integer, Integer> )SessionUtil.getInstance().getValue(req, "listItems");
-		
+		System.out.println(listItems.size());
 		boolean isLogin= false;
 		int customerID =0 ;
 		int orderID =0 ;
 		OrderModel orderModel = null;
-		if(userModel==null) {
+		if(listItems.size()==0) {
+			resData.put("status", 0);
+			resData.put("message", "Lỗi. Giỏ hàng trông");
+		} else if(userModel==null) {
 			model = RestUtil.of(req.getReader()).toModel(CustomerModel.class);
-			customerID = customerService.save(model);
-			model.setId(customerID);
-			orderModel = new OrderModel(new Date(),0,0,model.getHouseStreet(),model.getAreaId()
-					,model.getName(),model.getPhoneNumber(),model.getId());
-			orderID = orderService.save(orderModel);
-			orderModel.setId(orderID);
+			if(model.getAreaId()!=0&&model.getName().equals("")==false&&model.getEmail().equals("")==false
+					&&model.getPhoneNumber().equals("")==false&&model.getHouseStreet().equals("")==false) {
+				customerID = customerService.save(model);
+				model.setId(customerID);
+				orderModel = new OrderModel(new Date(),0,0,model.getHouseStreet(),model.getAreaId()
+						,model.getName(),model.getPhoneNumber(),model.getId());
+				orderID = orderService.save(orderModel);
+				orderModel.setId(orderID);				
+			} else {
+				model = null;
+				resData.put("status", 0);
+				resData.put("message", "Lỗi. Chưa nhập đầy đủ thông tin cá nhân người dùng");
+			}
 			
 		} else {
 			model = userModel.getCustomer();
@@ -76,24 +86,30 @@ public class OrderCartAPI extends HttpServlet{
 			orderID = orderService.save(orderModel);
 			orderModel.setId(orderID);
 		}
-		final int orderIDFinal = orderID;
-		List<OrderDetailsModel> list = listItems.entrySet().parallelStream().map(e->{
-			OrderDetailsModel oderDetailModel = new OrderDetailsModel(e.getKey(), e.getValue(),orderIDFinal);
-			return  oderDetailModel;
-		}).collect(Collectors.toList());
-		orderDetailsService.saveMulti(list);
-		if(RestUtil.message.toString().startsWith("45000")) {
-			orderService.delete(orderModel);
-			if(!isLogin) {
-				customerService.delete(model);				
-			}
-			resData.put("status", 0);
-			resData.put("message", RestUtil.message.toString().substring(6)
-					.replaceAll("[\n]", "").trim().replaceAll("\\s+", " "));
-			RestUtil.message.delete(0, RestUtil.message.length());
-		} else {
-			resData.put("status", 1);
+		
+		if(model!=null) {
+			
+			final int orderIDFinal = orderID;
+			List<OrderDetailsModel> list = listItems.entrySet().parallelStream().map(e->{
+				OrderDetailsModel oderDetailModel = new OrderDetailsModel(e.getKey(), e.getValue(),orderIDFinal);
+				return  oderDetailModel;
+			}).collect(Collectors.toList());
+			orderDetailsService.saveMulti(list);
+			if(RestUtil.message.toString().startsWith("45000")) {
+				orderService.delete(orderModel);
+				if(!isLogin) {
+					customerService.delete(model);				
+				}
+				resData.put("status", 0);
+				resData.put("message", RestUtil.message.toString().substring(6)
+						.replaceAll("[\n]", "").trim().replaceAll("\\s+", " "));
+				RestUtil.message.delete(0, RestUtil.message.length());
+			} else {
+				resData.put("status", 1);
+			}		
 		}
-		mapper.writeValue(resp.getOutputStream(), resData);
+			
+		 
+		mapper.writeValue(resp.getOutputStream(), resData);	
 	}	
 }
